Commit removes responsibility for reacting to device resizes from the side menu and pushes it upstream by allowing a prop to be passed in which has the same layout event related data.
This change will mean that we dont have nested layout events firing on children which testing has revealed can have inconsistent event data. We dont intend to contribute these fixes to the master repo for this library as its unfortunately not actively maintained.

We intend to eventually replace this drawer with a web equiv. E.g. material UI version, and source a different drawer for native use case.